data type in julia

+++ Triangles with equal side sums.jl
@@ -4,11 +4,11 @@ bottom=6
 scope=top-bottom+1
 timestart=time();
 figures=Array{Int}(undef,scope)
-for side=bottom:top
+for side::Int8=bottom:top
     figures[side-bottom+1]=0
-    for i=0:floor(side/3)
-        for j=i+1:floor((side-i)/2)
-            for k=j+1:side-i-j-1
+    for i::Int8=0:floor(side/3)
+        for j::Int8=i+1:floor((side-i)/2)
+            for k::Int8=j+1:side-i-j-1
                 #local corner=i+j+k
                 local a=side-i-j
                 local b=side-j-k
@@ -16,6 +16,7 @@ for side=bottom:top
                 if a!=i&&a!=j&&a!=k&&b!=i&&b!=j&&b!=k&&c!=i&&c!=j&&c!=k
                 #if k-i!=side-corner&&k-j!=side-corner&&j-i!=side-corner
                     figures[side-bottom+1]+=1
+                    #println(i,j,k,a,b,c)
                 end
             end  
         end
